+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1434,7 +1434,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-10T00:46:43.681Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-10T09:05:43.984Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1502,7 +1502,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-10T00:46:43.680Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-10T09:05:43.983Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1570,7 +1570,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-10T00:46:43.679Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-10T09:05:43.982Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1665,7 +1665,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-10T00:46:29.786Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-10T09:05:30.137Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")