+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH2G7KVK38N0RN8CBBXRH233"
+        "01KH3CSADTT2PZZR23J6A9BEKE"
     )
     r"""The unique identifier of the data source"""