+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH2G7KVTYF375VSNEX2A8GBW"
+        "01KH3CSAE4Q43QZFB2SQJGXAW8"
     )
     r"""The unique identifier of the data source"""