+++ packages/orq-rc/RELEASES.md
@@ -5429,3 +5429,13 @@ Based on:
 - [python v4.3.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.42 - packages/orq-rc
+
+## 2026-02-10 09:08:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.3 (2.814.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.43 - packages/orq-rc